♻️ refactor route setup
