+++ projects/01-NFT-Fi-Twitter/code/src/popup/main.ts
@@ -1,5 +1,5 @@
 import { createApp } from 'vue'
-import App from './Popup.vue'
+import App from './App.vue'
 import { setupApp } from '~/logic/common-setup'
 import '../styles'
